[Issue-10] Support Single Targets and Native-Grouped Targets
[Summary]
Some dependencies are only applicable to 1 target, and some apply to all
native targets (except WASM, which almost no one supports yet). It would
be nice to add in some target-specific dependencies, especially when
they relate closely to a shared dependency (Ktor and SQLDelight are
examples).

[Fix]
Create a "Single Target Libraries" and a "Native Target Libraries"
section. Single Target libraries are available as long as there is a
target selected that supports it. Native Target Libraries are available
as long as there is at least one native target it builds for (excluding
WASM).

[Testing]
- `./gradlew check`
- manual testing